Renamed ra to be remote_assistance since it's been bugging me what it stands for. Renamed files to be consistently named with underscores. Deleted ra_settings file as only function is much better suited to _db. Condensed the namespace where possible.
