+++ tp/src/plugins/remote_assistance/api/rd_session.py
@@ -1,22 +1,17 @@
 import json
 
 from vFense.core.api.base import BaseHandler
-from vFense.core.permissions._constants import *
-from vFense.core.permissions.decorators import check_permissions, authenticated_request
 
-from vFense.plugins import ra
-from vFense.plugins.ra import creator
+from vFense.plugins.remote_assistance import creator
 
 
 class RDSession(BaseHandler):
 
-    @authenticated_request
-    @check_permissions(Permissions.REMOTE_ASSISTANCE)
-    def post(self):
+    def post(self, agent_id=None):
 
         current_user = self.get_current_user()
-        body = json.loads(self.request.body)
-        agent_id = body.get('agent_id')
+        #body = json.loads(self.request.body)
+        #agent_id = body.get('agent_id')
 
         results = creator.new_rd_session(
             agent_id=agent_id,
@@ -26,13 +21,11 @@ class RDSession(BaseHandler):
         self.set_header('Content-Type', 'application/json')
         self.write(json.dumps(results, indent=4))
 
-    @authenticated_request
-    @check_permissions(Permissions.REMOTE_ASSISTANCE)
-    def delete(self):
+    def delete(self, agent_id=None):
 
         current_user = self.get_current_user()
-        body = json.loads(self.request.body)
-        agent_id = body.get('agent_id')
+        #body = json.loads(self.request.body)
+        #agent_id = body.get('agent_id')
 
         results = creator.terminate_rd_session(
             agent_id=agent_id,
