+++ tp/src/plugins/remote_assistance/api/rd_settings.py
@@ -2,8 +2,8 @@ import json
 
 from vFense.core.api.base import BaseHandler
 
-from vFense.plugins import ra
-from vFense.plugins.ra import creator
+from vFense.plugins import remote_assistance
+from vFense.plugins.remote_assistance._db import save_rd_password
 
 
 class RDPassword(BaseHandler):
@@ -14,7 +14,7 @@ class RDPassword(BaseHandler):
         body = json.loads(self.request.body)
         password = body.get('password')
 
-        results = creator.save_rd_password(
+        results = save_rd_password(
             password=password,
             user=current_user
         )
