+++ tp/src/plugins/remote_assistance/api/remote_desktop.py
@@ -1,17 +1,22 @@
 import json
 
 from vFense.core.api.base import BaseHandler
+from vFense.core.permissions._constants import *
+from vFense.core.permissions.decorators import check_permissions, authenticated_request
 
-from vFense.plugins.ra import creator
+from vFense.plugins import remote_assistance
+from vFense.plugins.remote_assistance import creator
 
 
 class RDSession(BaseHandler):
 
-    def post(self, agent_id=None):
+    @authenticated_request
+    @check_permissions(Permissions.REMOTE_ASSISTANCE)
+    def post(self):
 
         current_user = self.get_current_user()
-        #body = json.loads(self.request.body)
-        #agent_id = body.get('agent_id')
+        body = json.loads(self.request.body)
+        agent_id = body.get('agent_id')
 
         results = creator.new_rd_session(
             agent_id=agent_id,
@@ -21,11 +26,13 @@ class RDSession(BaseHandler):
         self.set_header('Content-Type', 'application/json')
         self.write(json.dumps(results, indent=4))
 
-    def delete(self, agent_id=None):
+    @authenticated_request
+    @check_permissions(Permissions.REMOTE_ASSISTANCE)
+    def delete(self):
 
         current_user = self.get_current_user()
-        #body = json.loads(self.request.body)
-        #agent_id = body.get('agent_id')
+        body = json.loads(self.request.body)
+        agent_id = body.get('agent_id')
 
         results = creator.terminate_rd_session(
             agent_id=agent_id,
