+++ tp/src/plugins/remote_assistance/novnc/websocket.py
