+++ tp/src/plugins/remote_assistance/novnc/websockify.py
